requirements.txt
decorator==4.3.0
ffmpeg-python==0.1.12
Flask==0.10
Flask-Bootstrap==3.2.0.2
future==0.16.0
imageio==2.3.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
moviepy==0.2.3.5
numpy==1.14.5
Pillow==2.5.1
simplejson==3.6.0
tqdm==4.23.4
Werkzeug==0.14.1
